% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlregVL.R
\name{add2ggDT}
\alias{add2ggDT}
\title{Add a geom, a theme, etc. to ggplots in a data.table}
\usage{
add2ggDT(dataT, gg2Str, graphCol, byC)
}
\arguments{
\item{dataT}{The data.table containing the ggplots}

\item{gg2Str}{String containing the unevaluated object to add to the ggplots. Can contain dataT column names.}

\item{graphCol}{name of the column containing the ggplot objects}

\item{byC}{Column to do the by.}
}
\value{
A data.table whose graphCol has been updated.
}
\description{
Add a geom, a theme, etc. to ggplots in a data.table
}

