% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlregVL.R
\name{doPsyFit}
\alias{doPsyFit}
\title{Applies a nonlinear regression on a specific dataset.}
\usage{
doPsyFit(xx, pp, type = "CG", getPred, residFun, fnJacob, fnStart, control)
}
\arguments{
\item{xx}{quantity on the x axis (e.g. duration). Must be a vector of numerics}

\item{pp}{quantity on the y axis (e.g. proportion of long responses). Must be a vector of numerics AND of the same length as xx.}

\item{type}{string expliciting which of three available fitting functions will be used. "CG" for cumulative Gaussian, "LN" for lognormal and "PL" for pseudologistic.}

\item{getPred}{Function that returns the predicted value.}

\item{residFun}{Function that returns the residuals.}

\item{fnJacob}{Function that returns the jacobian of the function to be fitted.}

\item{fnStart}{Function that returns the starting parameters.}

\item{control}{a control list for the non linear regression. See \code{\link[minpack.lm]{nls.lm.control}}}
}
\value{
a \code{\link{psyFitted-class}} object containing the fitted data.
}
\description{
Applies a nonlinear regression on a specific dataset.
}
\examples{
data("exampleData",package="nlregVL.R")
setkey(exampleData,Sub)
xx<-exampleData[J("TP01")][,Duration]
nbb1<-exampleData[J("TP01"),Nb1]
nbb3<-exampleData[J("TP01"),Nb3]
pp<-nbb3/(nbb1+nbb3)
result<-nlregVL::doPsyFit(xx = xx,pp = pp)
}

