% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlregVL.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example data}
\format{data.table with 105 observation of 4 variables.}
\usage{
data(exampleData)
}
\description{
A dataset used for examples
\itemize{
\item Duration. Duration in ms.
\item Sub. Participant code.
\item Nb1. Number of short responses.
\item Nb3. Number of long responses.
}
}

