% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlregVL.R
\name{fitLSDT}
\alias{fitLSDT}
\title{Run the non-linear regression on a data table}
\usage{
fitLSDT(dataF, cD, cNbS, cNbL, cBy, type = "CG", control, residFun, getPred,
  fnJacob, fnStart, rpsyF = T)
}
\arguments{
\item{dataF}{The data table containing the data to be fitted}

\item{cD}{name (or position) of the column containing the x-axis data. In the case of a temporal bisection task, this corresponds to the name (or position) of the column containing the duration.}

\item{cNbS}{name (or position) of the column containing the number of "short" responses.}

\item{cBy}{name (or position) of the column(s) to group by. If left empty, it will take all other columns.}

\item{type}{string expliciting which of three available fitting functions will be used. "CG" for cumulative Gaussian, "LN" for lognormal and "PL" for pseudologistic.}

\item{control}{a control list for the non linear regression. See \code{\link[minpack.lm]{nls.lm.control}}}

\item{residFun}{Function that returns the residuals.}

\item{getPred}{Function that returns the predicted value.}

\item{fnJacob}{Function that returns the jacobian of the function to be fitted.}

\item{fnStart}{Function that returns the starting parameters.}

\item{rpsyF}{Return psyF object as a column?}

\item{CNbL}{name (or position) of the column containing the number of "long" responses.}
}
\value{
A data.table with the parameter estimates, their confidence interval, the R square statistic and the Chi square statistic.
}
\description{
Run the non-linear regression on a data table
}
\examples{
data("exampleData",package="nlregVL")
result<-nlregVL::fitLSDT(dataF = exampleData,cD = "Duration",cNbS = "Nb1",cNbL = "Nb3",cBy = "Sub")

}
\seealso{
\code{\link{psyFitted-class}}
}

