% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlregVL.R
\name{fitslikDT}
\alias{fitslikDT}
\title{Maximum likelihood fits for psychometric functions}
\usage{
fitslikDT(dataF, cD, cNbS, cNbL, cBy, type = "CG", getPred, loglikfn,
  loglikgr, loglikgr2, fnStart, method = "BFGS", maxN = 100,
  parstart.allow = F, allow.print = F, ctrlL = list(fnscale = -1, reltol =
  .Machine$double.eps))
}
\arguments{
\item{dataF}{The data table containing the data to be fitted}

\item{cD}{name (or position) of the column containing the x-axis data. In the case of a temporal bisection task, this corresponds to the name (or position) of the column containing the duration.}

\item{cNbS}{name (or position) of the column containing the number of "short" responses.}

\item{cBy}{name (or position) of the column(s) to group by. If left empty, it will take all other columns.}

\item{type}{string expliciting which of three available fitting functions will be used. "CG" for cumulative Gaussian, "LN" for lognormal and "PL" for pseudologistic.}

\item{getPred}{Function that returns the predicted value.}

\item{loglikfn}{Only used if type is not specified. A function that computes the loglikelihood. It must take the following arguments: par (a vector of numerical values), Dur (a vector of durations), NbS (vector of the number of short respnses), NbL (vector of the number of short respnses). It must return a numerical vector of size 1.}

\item{loglikgr}{Only used if type is not specified. A function that computes the gradient of the loglikelihood. It must take the following arguments: par (a vector of numerical values), Dur (a vector of durations), NbS (vector of the number of short respnses), NbL (vector of the number of short respnses). It must return a numerical vector of the same size as par.}

\item{loglikgr2}{Only used if type is not specified. A function that computes the second order gradient of the loglikelihood. It must take the following arguments: par (a vector of numerical values), Dur (a vector of durations), NbS (vector of the number of short respnses), NbL (vector of the number of short respnses). It must return a numerical vector of the same size as par.}

\item{fnStart}{A function that calculates the starting values for the fitting process. It must take a single argument (Dur).}

\item{method}{See \code{\link[stats]{optim}}}

\item{maxN}{Maximum number of iterations.}

\item{parstart.allow}{If the starting parameters are bad, allows the algorithm to find new ones.}

\item{allow.print}{Allows the printing of the number of iterations of the fits.}

\item{ctrlL}{control list for optim. See \code{\link[stats]{optim}}}
}
\value{
A data.table containing the parameter estimates, the standard error of those estimates, the log likelihood as well as an information code (see \code{\link[stats]{optim}})
}
\description{
Maximum likelihood fits for psychometric functions
}

