% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlregVL.R
\name{newpsyFitset}
\alias{newpsyFitset}
\title{Creates an instance of a psyFitset object}
\usage{
newpsyFitset(type = "CG", xx, pp, getPred, residFun, fnJacob, fnStart)
}
\arguments{
\item{type}{string expliciting which of three available fitting functions will be used. "CG" for cumulative Gaussian, "LN" for lognormal and "PL" for pseudologistic.}

\item{xx}{quantity on the x axis (e.g. duration). Must be a vector of numerics}

\item{pp}{quantity on the y axis (e.g. proportion of long responses). Must be a vector of numerics AND of the same length as xx.}

\item{getPred}{Function that returns the predicted value.}

\item{residFun}{Function that returns the residuals.}

\item{fnJacob}{Function that returns the jacobian of the function to be fitted.}

\item{fnStart}{Function that returns the starting parameters.}
}
\value{
A psyFitset object for furture fitting.
}
\description{
Creates an instance of a psyFitset object
}
\examples{

data("exampleData",package="nlregVL")
setkey(exampleData,Sub)
setkey(exampleData,Sub)
xx<-exampleData[J("TP01")][,Duration]
nbb1<-exampleData[J("TP01"),Nb1]
nbb3<-exampleData[J("TP01"),Nb3]
pp<-nbb3/(nbb1+nbb3)
result<-nlregVL::newpsyFitset(xx = xx,pp = pp)

}
\seealso{
\code{\link{psyFitset-class}}
}

