% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlregVL.R
\docType{class}
\name{psyFitset-class}
\alias{psyFitset-class}
\title{Settings for the fitting process}
\description{
Settings for the fitting process
}
\section{Slots}{

\describe{
\item{\code{xx}}{quantity on the x axis (e.g. duration). Must be a vector of numerics}

\item{\code{pp}}{quantity on the y axis (e.g. proportion of long responses). Must be a vector of numerics AND of the same length as xx.}

\item{\code{getPred}}{Function that returns the predicted value.}

\item{\code{residFun}}{Function that returns the residuals.}

\item{\code{fnJacob}}{Function that returns the jacobian of the function to be fitted.}

\item{\code{fnStart}}{Function that returns the starting parameters.}
}}

