% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlregVL.R
\docType{class}
\name{psyFitted-class}
\alias{psyFitted-class}
\title{An S4 class containing the results of a non-linear regression by the levenberg-Marquardt algorithm}
\description{
An S4 class containing the results of a non-linear regression by the levenberg-Marquardt algorithm
}
\section{Slots}{

\describe{
\item{\code{psyF}}{psyFitset object containing the settings for the regression}

\item{\code{pars}}{list of the estimated parameters.}

\item{\code{Chi}}{Chi square statistic.}

\item{\code{ci}}{matrix containing the confidence intervals for the estimated parameters.}

\item{\code{R2}}{R square statistic.}

\item{\code{info}}{information. See \code{\link[minpack.lm]{nls.lm}}}
}}

